+++ backend/src/main/java/com/pigma/harusari/common/auth/controller/AuthController.java
@@ -66,22 +66,20 @@ public class AuthController {
     /* refreshToken 쿠키 생성 */
     private ResponseCookie createRefreshTokenCookie(String refreshToken) {
         return ResponseCookie.from("refreshToken", refreshToken)
-                .httpOnly(true)                     // HttpOnly 속성 설정 (JavaScript 에서 접근 불가)
-                // .secure(true)                    // HTTPS 환경일 때만 전송 (운영 환경에서 활성화 권장)
-                .path("/")                          // 쿠키 범위 : 전체 경로
-                .maxAge(Duration.ofDays(7))         // 쿠키 만료 기간 : 7일
-                .sameSite("Strict")                 // CSRF 공격 방어를 위한 SameSite 설정
+                .httpOnly(true)
+                .path("/")
+                .maxAge(Duration.ofDays(7))
+                .sameSite("Strict")
                 .build();
     }
 
     /* 쿠키 삭제용 설정 : 빈 값 + maxAge=0 으로 즉시 만료시켜 브라우저에서 삭제 */
     private ResponseCookie createDeleteRefreshTokenCookie() {
         return ResponseCookie.from("refreshToken", "")
-                .httpOnly(true)     // HttpOnly 유지
-                // .secure(true)    // HTTPS 환경에서만 사용 시 주석 해제
-                .path("/")          // 동일 path 범위
-                .maxAge(0)          // 즉시 만료
-                .sameSite("Strict") // SameSite 유지
+                .httpOnly(true)
+                .path("/")
+                .maxAge(0)
+                .sameSite("Strict")
                 .build();
     }
 }
